Splits `files` into `locate` and `content`.
