+++ src/utils/dependencies.js
@@ -1,6 +1,6 @@
-const {pkg} = require('./files');
+const {packageJson} = require('./files/contents');
 
-const find = (dependency) => pkg.dependencies?.[dependency] || pkg.devDependencies?.[dependency];
+const find = (dependency) => packageJson.dependencies?.[dependency] || packageJson.devDependencies?.[dependency];
 
 module.exports = {
     electron: find('electron'),
