--- src/utils/files.js
@@ -1,24 +1,0 @@
-const {join} = require('path');
-const {cwd} = require('process');
-
-const {existsSync, readJsonSync} = require('fs-extra');
-
-const locate = (paths) => {
-    for (const path of paths) {
-        const absolutePath = join(cwd(), path);
-
-        if (existsSync(absolutePath)) {
-            return absolutePath;
-        }
-    }
-
-    return undefined;
-};
-const pkg = readJsonSync(join(cwd(), 'package.json'), {throws: false}) || {};
-const tsconfig = readJsonSync(join(cwd(), 'tsconfig.json'), {throws: false}) || {};
-
-module.exports = {
-    locate,
-    pkg,
-    tsconfig,
-};
